2019_10_21 | django_auth

+++ 03_django/03_django_form/articles/views.py
@@ -1,17 +1,26 @@
 from IPython import embed
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
-from django.http import Http404
+from django.http import Http404, HttpResponse
 from django.views.decorators.http import require_POST
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
 
 # Create your views here.
 def index(request):
+    # session 에 visits_num 키로 접근해 값을 가져온다.
+    # 기본적으로 존재하지 않는 키이기 때문에 키가 없다면(방문한적이 없다면) 0 값을 가져오도록 한다.
+    visits_num = request.session.get('visits_num', 0)
+    # 그리고 가져온 값을 session 에 visits_num 에 매번 1씩 증가한 값으로 할당한다.(유저의 다음 방문을 위해)
+    request.session['visits_num'] = visits_num + 1
+    # session data 안에 있는 새로운 정보를 수정했다면 django 는 수정한 사실을 알아채지 못하기 때문에 다음과 같이 설정.
+    request.session.modified = True
+
     articles = Article.objects.all()
-    context = {'articles': articles,}
+    context = {'articles': articles,'visits_num': visits_num}
     return render(request, 'articles/index.html', context)
 
-
+@login_required
 def create(request):
     if request.method == 'POST':
         # form 인스턴스를 생성하고 요청에 의한 데이터를 인자로 받는다. (binding)
@@ -33,18 +42,20 @@ def create(request):
 def detail(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     comments = article.comment_set.all() # article 의 모든 댓글
-    comment_from = CommentForm() # 댓글 폼
-    context = {'article': article, 'comment_form': comment_from, 'comments': comments,}
+    comment_form = CommentForm() # 댓글 폼
+    context = {'article': article, 'comment_form': comment_form, 'comments': comments,}
     return render(request, 'articles/detail.html', context)
+
 
 @require_POST
 def delete(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        article.delete()
     return redirect('articles:index')
 
 
-
+@login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     if request.method == 'POST':
@@ -57,18 +68,23 @@ def update(request, article_pk):
     context = {'form': form, 'article': article,}
     return render(request, 'articles/form.html', context)
 
+
 @require_POST
-def comment_create(request, article_pk):
-    comment_form = CommentForm(request.POST)
-    if comment_form.is_valid():
-        # 객체를 Create 하지만, DB에 레코드는 작성하지 않는다.
-        comment = comment_form.save(commit=False)
-        comment.article_id = article_pk
-        comment.save() 
+def comments_create(request, article_pk):
+    if request.user.is_authenticated:
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            # 객체를 Create 하지만, db 에 레코드는 작성하지 않는다.
+            comment = comment_form.save(commit=False)
+            comment.article_id = article_pk
+            comment.save()
     return redirect('articles:detail', article_pk)
+
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
-    comment = get_object_or_404(Comment, pk=commpent_pk)
-    comment.delete()
-    return redirect('articlesLdetail', article_pk)
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        comment.delete()
+        return redirect('articles:detail', article_pk)
+    return HttpResponse('You ar Unauthorized', status=401)
